Bug 2035514 - Use HttpTrafficCategory for PHttpTransaction::Init. r=necko-reviewers,kershaw,timhuang
Differential Revision: https://phabricator.services.mozilla.com/D297040

UltraBlame original commit: 39d124eaa9a3b391ddd43a299afe57d23233005f

+++ netwerk/protocol/http/HttpTransactionParent.cpp
@@ -150,12 +150,11 @@ nsresult HttpTransactionParent::Init(
   
   
   if (!SendInit(caps, infoArgs, *requestHead, ipcStream, requestContentLength,
-                requestBodyHasHeaders, browserId,
-                static_cast<uint8_t>(trafficCategory), requestContextID,
-                classOfService, initialRwin, responseTimeoutEnabled, mChannelId,
-                !!mTransactionObserver, throttleQueue, mIsDocumentLoad,
-                aParentIpAddressSpace, aLnaPermissionStatus, mRedirectStart,
-                mRedirectEnd)) {
+                requestBodyHasHeaders, browserId, trafficCategory,
+                requestContextID, classOfService, initialRwin,
+                responseTimeoutEnabled, mChannelId, !!mTransactionObserver,
+                throttleQueue, mIsDocumentLoad, aParentIpAddressSpace,
+                aLnaPermissionStatus, mRedirectStart, mRedirectEnd)) {
     return NS_ERROR_FAILURE;
   }
